feat: manage choir collection lending
