+++ choir-app-frontend/src/app/features/choir-management/manage-choir/manage-choir.component.scss
@@ -96,3 +96,7 @@ mat-card-content form {
     background: rgba(0, 0, 0, 0.5);
   }
 }
+
+.library-icon {
+  margin-left: 8px;
+}
